% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrated.R
\name{integrated_ld_dat}
\alias{integrated_ld_dat}
\title{Integrated data
Create data list for laplaces demon for integrated model}
\usage{
integrated_ld_dat(ndat, sdat, pdat)
}
\arguments{
\item{ndat}{Data frame for the naive model}

\item{sdat}{Data frame for the survival model}

\item{pdat}{Data frame for population model}
}
\description{
Integrated data
Create data list for laplaces demon for integrated model
}
