% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrated.R
\name{integrated_lp}
\alias{integrated_lp}
\title{Integrated model
Log probability integrating naive and survival and population models}
\usage{
integrated_lp(parm, Data)
}
\arguments{
\item{parm}{Parameter list}

\item{Data}{Data list}
}
\description{
Integrated model
Log probability integrating naive and survival and population models
}
