% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive.R
\name{naive_model_stan}
\alias{naive_model_stan}
\title{Naive Distribution Model}
\usage{
naive_model_stan(iter = 1000, chains = 1, repeatable = FALSE,
  print_stan_messages = FALSE)
}
\arguments{
\item{iter}{number of hmc iterations}

\item{chains}{number of hmc chains}

\item{repeatable}{boolean, should we use a fixed seed?}

\item{print_stan_messages}{boolean, show the (copius) stan messages?}
}
\value{
list of stan models, one per species
}
\description{
Naive distribution model from seedling metamodel, impemented in rstan.
Created by Paige E. Copenhaver-Parry  28 January 2017
}
