% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrated.R
\name{param_index}
\alias{param_index}
\title{Utility function to find the index position in a LD parameter object of a named parameter}
\usage{
param_index(dat, parm)
}
\arguments{
\item{dat}{Data list to search}

\item{parm}{Parameter name}
}
\description{
Utility function to find the index position in a LD parameter object of a named parameter
}
