% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{pop_ld_dat}
\alias{pop_ld_dat}
\title{Population model data}
\usage{
pop_ld_dat(dat)
}
\arguments{
\item{dat}{Data frame for the survival model}
}
\description{
Create data list for laplaces demon for survival model
}
