% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruitment.R
\name{recruitment_ld_dat}
\alias{recruitment_ld_dat}
\title{Recruitment submodel data}
\usage{
recruitment_ld_dat(dat)
}
\arguments{
\item{dat}{Data frame for the recruitment model}
}
\description{
Create data list for laplaces demon for recruitment model
}
