% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruitment.R
\name{recruitment_lp}
\alias{recruitment_lp}
\title{Recruitment submodel from seedling metamodel
Created by Paige E. Copenhaver-Parry  28 January 2017
Modified 6 Dec 2017}
\usage{
recruitment_lp(parm, Data, nested = FALSE)
}
\arguments{
\item{parm}{Parameter list}

\item{Data}{Data list}

\item{nested}{Boolean, is the model nested within another likelihood?}
}
\description{
Recruitment submodel from seedling metamodel
Created by Paige E. Copenhaver-Parry  28 January 2017
Modified 6 Dec 2017
}
\details{
If \code{nested} is \code{TRUE}, only the log likelihood and 
log posterior are returned, along with the model predictions. Otherwise
a list conforming to LaplacesDemon model functions is returned
}
