% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{survival_ld_dat}
\alias{survival_ld_dat}
\title{Survival model data}
\usage{
survival_ld_dat(dat)
}
\arguments{
\item{dat}{Data frame for the survival model}
}
\description{
Create data list for laplaces demon for survival model
}
