% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{survival_lp}
\alias{survival_lp}
\title{Survival model}
\usage{
survival_lp(parm, Data, nested = FALSE)
}
\arguments{
\item{parm}{Parameter list}

\item{Data}{Data list}

\item{nested}{Boolean, is the model nested within another likelihood?}
}
\description{
Survival model from seedling metamodel
Created by Paige E. Copenhaver-Parry  28 January 2017
}
\details{
If \code{nested} is \code{TRUE}, only the log likelihood and 
log posterior are returned, along with the model predictions. Otherwise
a list conforming to LaplacesDemon model functions is returned
}
