% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrated.R
\name{tss}
\alias{tss}
\title{Compute the true skill statistic}
\usage{
tss(thresh, y, pr)
}
\arguments{
\item{y}{Data points; observed 1s and 0s}

\item{pr}{Predicted probabilities}

\item{thres}{The threshold to apply if probabilities (pr) are specified}
}
\description{
Compute the true skill statistic
}
